Add custom error classes for invalid characters and unclosed strings; refactor error handling in `Escaner`
