+++ src/main/java/analizadorLexico/Errores/DoubleInvalidoError.java
@@ -1,0 +1,6 @@
+package analizadorLexico.Errores;
+
+public class DoubleInvalidoError extends ErrorLex {
+    public DoubleInvalidoError(int line, int column, String lexema){
+        super (line, column, "DOUBLE INVALIDO ", lexema);}
+}
