+++ src/main/java/analizadorLexico/Escaner.java
@@ -1,5 +1,8 @@
 package analizadorLexico;
 
+
+import analizadorLexico.Errores.*;
+
 import java.io.IOException;
 import java.util.HashMap;
 import java.util.Map;
@@ -236,7 +239,7 @@ public class Escaner {
                                 column = 0;
                             }
                             if (isAtEnd()) {
-                                throw new ErrorLex(line, column, "NO CIERRA COMENTARIO MULTILINEA", String.valueOf(c));
+                                throw new ComentarioSinCerrarError(line, column, String.valueOf(c));
                                 //throw new IOException("CARACTER INVALIDO en línea " + line + ", columna " + column);
                             }
 
@@ -255,7 +258,7 @@ public class Escaner {
                     //Pasamos de largo el comentario
                     return addToken(OR);
                 } else {
-                    throw new ErrorLex(line, column, "CARACTER INVALIDO", String.valueOf(c));
+                    throw new CaracterInvalidoError(line, column, String.valueOf(c));
                     //throw new IOException("CARACTER INVALIDO en línea " + line + ", columna " + column);
                 }
 
@@ -265,7 +268,7 @@ public class Escaner {
                     //Pasamos de largo el comentario
                     return addToken(AND);
                 } else {
-                    throw new ErrorLex(line, column, "CARACTER INVALIDO", String.valueOf(c));
+                    throw new CaracterInvalidoError(line, column, String.valueOf(c));
                 }
                 //literales cadenas
             case '"':
@@ -287,7 +290,7 @@ public class Escaner {
             if (isAlpha(c)) {
                 return identifier(c);
             } else {
-                throw new ErrorLex(line, column, "CARACTER INVALIDO", String.valueOf(c));
+                throw new CaracterInvalidoError(line, column, String.valueOf(c));
                 //throw new IOException("CARACTER INVALIDO en línea " + line + ", columna " + column);
             }
 
@@ -393,7 +396,7 @@ public class Escaner {
             advance();
         }
         if (isAtEnd()) {
-            throw new ErrorLex(line, column, "STRING SIN TERMINAR", buffer.substring(start + 1, current - 1));
+            throw new StringSinCerrarError(line, column, buffer.substring(start + 1, current - 1));
         }
 
         advance();
@@ -450,12 +453,12 @@ public class Escaner {
                 isDouble = true;
             }else {
 
-                throw new ErrorLex(line, column, "DOUBLE INVALID", buffer.substring(start + 1, current - 1));
+                throw new DoubleInvalidoError(line, column, buffer.substring(start + 1, current - 1));
             }
         }
         if(isAlpha(look())) {
 
-            throw new ErrorLex(line, column, "CARACTER INVALIDO", buffer.substring(start + 1, current - 1));
+            throw new CaracterInvalidoError(line, column, buffer.substring(start + 1, current - 1));
         }
 
         while (isDigit(look())) advance();
