+++ src/main/java/analizadorLexico/Etapa1.java
@@ -1,4 +1,7 @@
 package analizadorLexico;
+
+import analizadorLexico.Errores.ErrorLex;
+
 import java.io.BufferedWriter;
 import java.io.FileWriter;
 import java.io.IOException;
